requirements.txt
attrs
click
deap
numpy
gym
scipy
torch
procgen
opencv-python
cma
matplotlib
pandas
py-cpuinfo
pytest
tensorflow
tensorboardX